Move torch-related functions from utils.py to nn.nn_utils.py
