+++ libmultilabel/nn/nn_utils.py
@@ -1,10 +1,37 @@
+import logging
+import os
+
+import numpy as np
 import pytorch_lightning as pl
+import torch
 from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 from pytorch_lightning.callbacks.model_checkpoint import ModelCheckpoint
 from pytorch_lightning.utilities.seed import seed_everything
 
 from ..nn import networks
 from ..nn.model import Model
+
+
+def init_device(use_cpu=False):
+    """Initialize device to CPU if `use_cpu` is set to True otherwise GPU.
+
+    Args:
+        use_cpu (bool, optional): Whether to use CPU or not. Defaults to False.
+
+    Returns:
+        torch.device: One of cuda or cpu.
+    """
+    if not use_cpu and torch.cuda.is_available():
+        # Set a debug environment variable CUBLAS_WORKSPACE_CONFIG to ":16:8" (may limit overall performance) or ":4096:8" (will increase library footprint in GPU memory by approximately 24MiB).
+        # https://docs.nvidia.com/cuda/cublas/index.html
+        os.environ['CUBLAS_WORKSPACE_CONFIG'] = ":4096:8"
+        device = torch.device('cuda')
+    else:
+        device = torch.device('cpu')
+        # https://github.com/pytorch/pytorch/issues/11201
+        torch.multiprocessing.set_sharing_strategy('file_system')
+    logging.info(f'Using device: {device}')
+    return device
 
 
 def init_model(model_name,
@@ -104,3 +131,20 @@ def init_trainer(checkpoint_dir,
                          max_epochs=epochs,
                          callbacks=[checkpoint_callback, earlystopping_callback])
     return trainer
+
+
+def set_seed(seed):
+    """Set seeds for numpy and pytorch.
+
+    Args:
+        seed (int): Random seed.
+    """
+
+    if seed is not None:
+        if seed >= 0:
+            seed_everything(seed=seed)
+            torch.use_deterministic_algorithms(True)
+            torch.backends.cudnn.benchmark = False
+        else:
+            logging.warning(
+                f'the random seed should be a non-negative integer')
