+++ torch_trainer.py
@@ -11,8 +11,8 @@ from pytorch_lightning.callbacks.model_checkpoint import ModelCheckpoint
 from libmultilabel.nn import data_utils
 from libmultilabel.nn import networks
 from libmultilabel.nn.model import Model
-from libmultilabel.nn.nn_utils import init_model, init_trainer
-from libmultilabel.utils import dump_log, init_device, set_seed
+from libmultilabel.nn.nn_utils import init_device, init_model, init_trainer, set_seed
+from libmultilabel.utils import dump_log
 
 
 class TorchTrainer:
